Reverted changes on get() function
My misunderstanding of this function introduced a breaking change, which @oliverfoster caught.

A better solution is to standardise the handling of undefined or assessment 'id' values which are empty strings.  Warning messages are now output when appropriate, but the code allows an empty assessment 'id', e.g. if there is only one assessment on the course.